# Compare bytes with bytes.Equal instead of reflect.DeepEqual