@bgaifullin 2017-05-02
I this this is covered by PR #394